Rename the component from square card to tilted card
